solved personal music on profile page

+++ server/schemas/resolvers.js
@@ -5,10 +5,8 @@ const { signToken } = require("../utils/auth");
 const fs = require('fs')
 const util = require('util')
 const unlinkFile = util.promisify(fs.unlink)
-
 const multer = require('multer')
 const upload = multer({ dest: 'uploads/' })
-const { uploadFile, getFileStream } = require('../utils/s3')
 
 const resolvers = {
   Query: {
@@ -29,7 +27,7 @@ const resolvers = {
         throw new Error("Error fetching posts");
       }
     },
-    
+
     users: async () => {
       const user = await User.find();
       return user;
@@ -47,6 +45,12 @@ const resolvers = {
       const user = await User.findOne({ username: args.username });
       const posts = await Post.find({ user: user._id }).populate("user");
       return posts
+    },
+    getUsersSongs: async (parent, args, context) => {
+      const user = await User.findOne({ username: args.username });
+      const songs = await Music.find({ user: user._id }).populate("user");
+      console.log('songs'+songs)
+      return songs;
     },
   },
   Mutation: {
@@ -83,9 +87,9 @@ const resolvers = {
         { username: args.username },
         {
           email: args.email,
-          firstName: args.firstName, 
+          firstName: args.firstName,
           lastName: args.lastName,
-          bio: args.bio 
+          bio: args.bio
         },
         { new: true }
       )
@@ -114,17 +118,16 @@ const resolvers = {
         try {
           const user = await User.findById(context.user._id);
           const newPost = new Post({
-            user: user._id, 
+            user: user._id,
             title,
             description,
             images,
-            profileImage,
           });
           const savedPost = await newPost.save();
-          
-          await User.findOneAndUpdate({_id:user._id},{$push:{posts:savedPost}},{new:true});
+
+          await User.findOneAndUpdate({ _id: user._id }, { $push: { posts: savedPost } }, { new: true });
           const populatedPost = await Post.findById(savedPost._id).populate('user').exec();
-    
+
           return populatedPost;
         } catch (error) {
           console.error(error);
@@ -136,7 +139,7 @@ const resolvers = {
     followUser: async (parent, args, context) => {
       if (context.user) {
         try {
-          
+
           const follower = await User.findById(context.user._id);
           const followee = await User.findOneAndUpdate(
             { username: args.username },
@@ -157,19 +160,6 @@ const resolvers = {
       };
       throw new Error('Authentication Error. Please sign in.');
     },
-    
-
-
-    
-
-    // uploadImage: async (parent, { file }, context) => {
-    //   const result = await uploadFile(file)
-    //   await unlinkFile(file.path)
-    //   console.log(result)
-
-    //   return {imagePath: `/images/${result.Key}`}
-    // },
-
     createComment: async (parent, { input }, context) => {
       if (context.user) {
         const { postId, content } = input;
@@ -190,38 +180,30 @@ const resolvers = {
       }
       throw new AuthenticationError("You need to be logged in to perform this action")
     },
-    // saveMusic: async (parent, { title, artist, url, coverart }) => {
-    //   console.log(title, artist, url, coverart);
-    //   try {
-    //     const music = new Music({ title, artist, url, coverart });
-    //     return await music.save();
-    //   } catch (error) {
-    //     console.error(error);
+    saveMusic: async (parent, { title, artist, url, coverart }, context) => {
+      if (context.user) {
+        try {
+          const user = await User.findById(context.user._id);
+          let music = await Music.findOne({ title });
+          if (music) {
 
-    //     throw new Error('Error creating music');
-    //   }
-    // },
-    saveMusic: async (parent, { title, artist, url, coverart },context) => {
-      try {
-        let music = await Music.findOne({ title });
-        if(music){
-       
-          await Music.findOneAndDelete({ title });
-        } else {
-         
-          music = new Music({ title, artist, url, coverart });
-          await music.save();
+            await Music.findOneAndDelete({ title });
+          } else {
+
+            music = new Music({ title, artist, url, coverart, user: user._id });
+            await music.save();
+          }
+
+          const savedUser = await User.findById(context.user._id);
+          savedUser.musics.push(music);
+          await savedUser.save();
+        } catch (error) {
+          console.error(error);
+          throw new Error('Error in saveMusic mutation');
         }
-        
-        const savedUser = await User.findById(context.user._id);
-        savedUser.musics.push(music);
-        await savedUser.save();
-      } catch (error) {
-        console.error(error);
-        throw new Error('Error in saveMusic mutation');
       }
     },
-    
+
     deleteMusic: async (parent, { title }) => {
       try {
         const music = await Music.findOne({ title });
@@ -235,7 +217,7 @@ const resolvers = {
         throw new Error('Error deleting music');
       }
     },
-    
+
     register: async (parent, { username, email, password, firstName, lastName }) => {
       const user = await User.create({ username, email, password, firstName, lastName });
       const token = signToken(user);
@@ -243,8 +225,7 @@ const resolvers = {
     },
 
   },
-};
+}
+  ;
 
 module.exports = resolvers
-
-
